requirements.txt
python-dotenv
requests
sendgrid
pytest